E2E: press the power button when a scenario fails
check() re-raises, so a failure between switching the machine off and the
press that follows skipped the press. The machine was left off, the
setting restore failed with it, and so did the runner's teardown and
every suite after it - none of which resemble the failure that caused it.
recover_if_off() now runs on the way out of both u64ctrl suites and is a
no-op when the machine is up.

Mains moves to machine_power.py beside PowerButton, since a second suite
needs it.
